Require explicit thousands separator (#7)
In https://github.com/mara/mara-google-sheet-downloader/pull/7 the problem came up that 1,000 needs to be interepted as 1000.0 and not as 1.000. With the old logic this wasn't possible. Now we explicitly expect a thousands separator which (per default) is ','. If you spreadsheet is e.g. German, you need to set the thousands separator in your column spec: '...f(thousands_separator=.)...'

Closes https://github.com/mara/mara-google-sheet-downloader/pull/7